:truck: Move GithubLink
